fix collection not set for vocab

+++ src/pages/vocab.py
@@ -5,7 +5,6 @@ from pages.utils import Db
 st.header("Create vocabulary collection")
 
 
-
 def set_data():
     st.session_state["vocab_data"] = Db.get_ids("vocab", [st.session_state["vocab_collection"]])
     st.session_state["stopwords_data"] = Db.get_ids("vocab", st.session_state["stopwords_collections"])
@@ -27,19 +26,19 @@ with st.sidebar:
 
 col1, col2 = st.columns(2, gap="large")
 with col1:
-    collection = st.session_state["vocab_collection"]
-    st.subheader(f"Vocabulary {collection}")
-    if st.session_state.get("vocab_data"):
-        select_all = st.checkbox("Select All", value=True)
-        with st.form('vocab_form'):
-            checkboxes = []
-            with st.container(height=500, border=False):
-                for word in st.session_state["vocab_data"]:
-                    checkboxes.append((st.checkbox(f"{word}", value=select_all), word))
-            if st.form_submit_button("Save", type="primary"):
-                db = Db.get_db("vocab", collection)
-                db.remove_ids([word for checked, word in checkboxes if not checked])
-                set_data()
+    if collection := st.session_state["vocab_collection"]:
+        st.subheader(f"Vocabulary {collection}")
+        if st.session_state.get("vocab_data"):
+            select_all = st.checkbox("Select All", value=True)
+            with st.form('vocab_form'):
+                checkboxes = []
+                with st.container(height=500, border=False):
+                    for word in st.session_state["vocab_data"]:
+                        checkboxes.append((st.checkbox(f"{word}", value=select_all), word))
+                if st.form_submit_button("Save", type="primary"):
+                    db = Db.get_db("vocab", collection)
+                    db.remove_ids([word for checked, word in checkboxes if not checked])
+                    set_data()
         # st.write(data)
 with col2:
     st.subheader("Jobs keywords")
